Fix AHIK simulation and re-enable distinct operands.

+++ src/s390/constants-s390.h
@@ -1650,7 +1650,7 @@ class RIEInstruction : Instruction {
     return Bits<SixByteInstr, uint32_t>(15, 8);
   }
   inline int I6Value() const {
-    return Bits<SixByteInstr, int32_t>(31, 16);
+    return static_cast<int32_t>(Bits<SixByteInstr, int16_t>(31, 16));
   }
   inline int size() const {
     return 6;
